+++ src/components/fontawesome-icons.jsx
@@ -6,8 +6,6 @@ import {
   faGlobeAmericas,
   faUserFriends,
   faPaperclip,
-  faBell,
-  faUser,
 } from '@fortawesome/free-solid-svg-icons';
 import { faComment, faHeart as faHeartO, faImage } from '@fortawesome/free-regular-svg-icons';
 import { faCommentPlus } from './fontawesome-custom-icons';
@@ -26,8 +24,6 @@ const preloadedIcons = [
   faUserFriends,
   faImage,
   faPaperclip,
-  faBell,
-  faUser,
 ];
 
 export const SVGSymbolDeclarations = memo(function SVGSymbolDeclarations() {