+++ webpack/rules.js
@@ -38,7 +38,7 @@ class RuleGenerator {
                 {
                   modules: false,
                   useBuiltIns: 'entry',
-                  corejs: '3.23',
+                  corejs: '3.25',
                 },
               ],
             ]),
@@ -77,7 +77,7 @@ class RuleGenerator {
       loader: 'babel-loader',
       options: {
         babelrc: false,
-        targets: { node: '12' },
+        targets: { node: '14' },
         presets: [
           [
             '@babel/react',
@@ -167,9 +167,10 @@ class RuleGenerator {
       },
       {
         test: /vazir-font[\\/]dist.*/,
-        loader: 'file-loader',
-        options: {
-          name: 'assets/fonts/[name].[ext]',
+        type: 'asset/resource',
+        dependency: { not: ['url'] },
+        generator: {
+          filename: 'assets/fonts/[name].[ext]',
         },
       },
     ];
@@ -180,9 +181,9 @@ class RuleGenerator {
 
     return {
       test: /photoswipe.+\.(png|svg|gif)$/,
-      loader: `file-loader`,
-      options: {
-        name: `assets/images/photoswipe/${fileName}`,
+      type: 'asset/resource',
+      generator: {
+        filename: `assets/images/photoswipe/${fileName}`,
       },
     };
   }